REF: Use `head_object` for .info
This should be more efficient for `.info` calls.
The response dict is almost the same. They use `ContentLength` instead
of `Size` and don't include the key in the response.

One question I had was on exception handling. To get the test suite to
pass, I had to catch botocore's `ParamvalidationError.` Otherwise things
like `s3.open('x', 'rb')` failed with a ParamvalidationError, since `'x'`
isn't a valid bucket name, while the tests were expecting an IOError or
OSError. Would be curious to hear what you think on this.

With this the pandas test-suite pretty much passes.
My changes are
[here](https://github.com/pydata/pandas/commit/e89e54a32e3b53764005da488215dd529b20ea6c)
Just a couple tests to fix up that were expecting different error
messages.

http://boto3.readthedocs.org/en/latest/reference/services/s3.html#S3.Client.head_object